Create dice role
The game will prompt you to roll a dice and add the roll to your score. You have a maximum of 3 rolls to reach a target score of 20. If you reach the target score within the given number of rolls, you win the game. Otherwise, the game ends, and you lose.